Discard deprecated lib
