+++ src/cloud/lib/export.ts
@@ -9,7 +9,7 @@ import rehypeRaw from 'rehype-raw'
 import rehypeSanitize from 'rehype-sanitize'
 import rehypeKatex from 'rehype-katex'
 import { prop } from 'ramda'
-import { downloadString } from './download'
+import { downloadString } from '../../design/lib/dom'
 import rehypeCodeMirror from '../../design/lib/codemirror/rehypeCodeMirror'
 import { SerializedDoc } from '../interfaces/db/doc'
 import { filenamify } from './utils/string'
