--- src/lib/db/consts.ts
@@ -1,4 +1,0 @@
-export const FOLDER_ID_PREFIX = 'folder:'
-export const NOTE_ID_PREFIX = 'note:'
-export const TAG_ID_PREFIX = 'tag:'
-export const ATTACHMENTS_ID = 'attachments'
