--- src/lib/db/patterns.ts
@@ -1,5 +1,0 @@
-import { NavResource } from '../v2/interfaces/resources'
-
-export function getResourceId(source: NavResource) {
-  return source.result._id
-}
