--- src/lib/keyboard.ts
@@ -1,35 +1,0 @@
-import { useEffect } from 'react'
-import { osName } from './platform'
-
-export const useGlobalKeyDownHandler = (
-  handler: (event: KeyboardEvent) => void
-) => {
-  return useEffect(() => {
-    window.addEventListener('keydown', handler)
-    return () => {
-      window.removeEventListener('keydown', handler)
-    }
-  }, [handler])
-}
-
-export function isWithGeneralCtrlKey(
-  event: KeyboardEvent | React.KeyboardEvent
-) {
-  switch (osName) {
-    case 'macos':
-      return event.metaKey
-    default:
-      return event.ctrlKey
-  }
-}
-
-export function isWithGeneralCtrlShiftKeys(
-  event: KeyboardEvent | React.KeyboardEvent
-) {
-  switch (osName) {
-    case 'macos':
-      return event.metaKey && event.shiftKey
-    default:
-      return event.ctrlKey && event.shiftKey
-  }
-}
