--- src/lib/localStorageKeys.ts
@@ -1,12 +1,0 @@
-export const preferencesKey = 'note.boostio.co:preferences'
-export const storageDataListKey = 'note.boostio.co:storageDataList'
-export const generalStatusKey = 'note.boostio.co:generalStatusKey'
-export const previewStyleKey = 'note.boostio.co:previewStyleKey'
-export const defaultStorageCreatedKey =
-  'note.boostio.co:defaultStorageCreatedKey'
-export const appModeChosenKey = 'note.boostio.co:appModeChosen'
-export const checkedFeaturesKey = 'note.boostio.co:checkedFeaturesKey'
-export const sidebarCollapseKey = 'note.boostio.co:sidebar'
-
-// Deprecated
-export const introDataKey = 'note.boostio.co:introDataKey'
